+++ recipes/preboot.rb
@@ -1,6 +1,0 @@
-#
-# Cookbook Name:: dop_base
-# Recipe:: preboot
-#
-
-include_recipe "dop_base::user"