+++ test/integration/default/serverspec/default_spec.rb
@@ -13,18 +13,20 @@ describe host('dop') do
   it { should be_resolvable.by('hosts') }
 end
 
-describe package('htop') do
-  it { should be_installed }
-end
-
 describe file('/etc/ssh/ssh_known_hosts') do
   it { should be_file }
   it { should be_mode 644 }
   it { should be_owned_by 'root' }
   it { should be_grouped_into 'root' }
-  its(:content) { should match /github.com/ }
 end
 
-describe package('nmap') do
-  it { should be_installed }
+%w(htop nmap siege python-pip libxml-xpath-perl ntp ntpdate).each do |pkg|
+  describe package(pkg) do
+    it { should be_installed }
+  end
+end
+
+describe service('ntp') do
+  it { should be_enabled }
+  it { should be_running }
 end