+++ test/integration/default/serverspec/dotdeb_spec.rb
@@ -9,7 +9,7 @@ RSpec.configure do |c|
   end
 end
 
-describe file('/etc/apt/sources.list.d/dotdeb.list') do
+describe file('/etc/apt/sources.list.d/dotdeb.list'), if: os[:family] == 'Debian' do
   it { should be_file }
   it { should be_mode 644 }
   it { should be_owned_by 'root' }