+++ test/integration/default/serverspec/ssh_spec.rb
@@ -16,6 +16,13 @@ describe file('/etc/ssh/sshd_config') do
   it { should be_grouped_into 'root' }
 end
 
+describe file('/root/.bashrc') do
+  it { should be_file }
+  it { should be_mode 644 }
+  it { should be_owned_by 'root' }
+  it { should be_grouped_into 'root' }
+end
+
 describe service('ssh') do
   it { should be_enabled }
   it { should be_running }